MTV-2858 | Add UDN IP address preservation
The UDN in the OCP 4.20.0 will have new feature for setting IP addresses
on the VMs. This will allow forklift to specify which IP address the VM
should have after migration. This way we will be able to preserve the IP
from original provider. This requires that the IP address is from the
same subnet.
Forklift will preserrve this with the `preserveStaticIPs` parameter. So
from the User Interface perspectvie no configuration will change, but
the flag will not only preserve static IPs but also set IP addresses in
the UDNs.

Notes:
- Right now we can not specify default gateway, DNS, secondary UDNs

Ref: https://issues.redhat.com/browse/MTV-2858

Signed-off-by: Martin Necas <mnecas@redhat.com>