storage-offload: Fix device discovery timeout
Issue:
When the device is mapped to the ESXi host, it can take some time to get
discovered by the kernel. Right now, we wait only if the rescan fails,
but in cases where the scan passes, we immediately check for the device.

Fix:
Always add timeout for the next device discovery try.

Signed-off-by: Martin Necas <mnecas@redhat.com>