Add SSHReadiness condition
Add SSHReadiness condition on the provider and plan level.
In the condition add instructions on how to properly enable ssh on the
ESXi.
A change to the restricted command code has been made to accommodate the
fact we do not have a datastore when we do the validation of the esxi,
and so can't run the python script (since it is uploaded to a specific
datastore)).

Resolves: https://issues.redhat.com/browse/MTV-3590
Resolves: https://issues.redhat.com/browse/MTV-3591

Co-authored with Cursor
Signed-off-by: borod108 <boris.od@gmail.com>