Move the EsxHost connection test to lib
After adding the previous libvirt support, the forklift-api image gained
an indirect dependency on libvirt. In order to break this dependency,
move the EsxHost TestConnection to a new lib module that won't pull in
other parts of the controller. This allows the forklift-api image to
avoid a dependency on libvirt.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>