build: Rename upstream ovirt-populator Containerfile
The ovirt-populator build directory has both `Containerfile` and
`Containerfile-upstream`. The `Containerfile` file is left over from
early on before the build was split between upstream and downstream
builds. It is currently not referenced in the makefile so it appears
to be unused.  For consistency with the other images where the
upstream build file is simply named `Containerfile`, delete the
unused containerfile and rename the `Containerfile-upstream` to
`Containerfile`.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>