MTV-3394 | storage-offload: fix the rescan handling
Fix the rescan logic so it fails in case the rescan succeeds, but the
device is still not detected.

Also, fix the device argument to be only the naa.XYZ (or any other world
wide name) of the device, and not the /path/to/naa.XYZ as the 'esxcli
storage core device list -d XXX' doesn't handle the path.

Signed-off-by: Roy Golan <rgolan@redhat.com>