storage-offload: Pure flasharray - fix the FC host lookup and matchup
Pure flasharray stores the wwn on the host in this form:
21:00:00:0E:1E:1E:83:D0
While the adapter on the ESX is reporting:
fc.2100000e1e1e83d0:2100000e1e1e83d2

This fix handles that difference when matching. If the matching fails
and we don't match any hosts then the Populate function returns with an
err.

Signed-off-by: Roy Golan <rgolan@redhat.com>