storage-offload: Pure flasharray - fix the FC host lookup and matchup
Pure flasharray stores the WWPN in the host's wwn field in this form:
21:00:00:0E:1E:1E:83:D0
While the adapter on the ESX is reporting the full WWN in the form of
fc.{WWNN}:{WWNPN} for example:
fc.2100000e1e1e83d0:2100000e1e1e83d2

This fix handles that difference when matching and will attempt to match
the WWPN part. If the matching fails the Populate function returns with an err.

Signed-off-by: Roy Golan <rgolan@redhat.com>