MTV-3090 | copy-offload: Add an indication that xcopy was used
Resolves: MTV-3090

Motivation
Copying using vmkfstools may fallback to host copy due to configuration errors on the
storage array or network issues between the ESX and the array. If and
when that happens or whenever XCOPY was actually in used we want to
know, so we can alert, analyze, troubleshoot and improve.

Modification
Use ESX tooling to detect the number of accelerated operations on a
target device and if its more than 0 then report it on the task status
data as {"xcopyUsed": true}

Result
popoulator pod can detect xcopy usage and report and publishes a new
metric along with the progress with the name:
"vsphere_xcopy_volume_populator_xcopy_used"

Signed-off-by: Roy Golan <rgolan@redhat.com>