MTV-3473 | Remove the fetch before plan update
Issue:
When the plan starts to reconcile it adds condition `The plan is EXECUTING`.
But when we try to update the new resrouce the old condition the statuis
gets removed.
So in the next reconcile the condintion is missing as we don't add it to
the condition.
```
if snapshot.HasAnyCondition(api.ConditionExecuting, api.ConditionSucceeded, api.ConditionFailed, api.ConditionCanceled) {
	return
}
r.Plan.Status.Migration.MarkReset()
r.Plan.Status.Migration.MarkStarted()
snapshot.SetCondition(
	libcnd.Condition{
		Type:     api.ConditionExecuting,
		Status:   True,
		Category: api.CategoryAdvisory,
		Message:  "The plan is EXECUTING.",
		Durable:  true,
	})
```
This causes the update to remove this condition and restart the whole
plan.

Fix:
Remove the fetch before plan update.

Signed-off-by: Martin Necas <mnecas@redhat.com>