OCP: Treat ref with no namespace as ambiguous
The OCP inventory finder currently tolerates
being given a ref with only an unqualified
name set, and it only returns a RefNotUnique
error if more than one namespace-scoped resource
with that name exists in the inventory. A name
without a namespace for a namespace-scoped
resource is inherently ambiguous and should
be rejected.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>