Only attempt to update Plan status in one place
The Plan status was getting updated before the call
to `execute()`, and then once during or after `execute()`
as well. This frequently results in conflicts during the
second attempt to update the status, and updating the status
after `execute()` returns with error can result in the Plan
entering an inconsistent state. (for instance, if the
migration runner returned with an error after the `Executing`
condition had been set, but before `begin()` had
completed successfully.)

Signed-off-by: Sam Lucidi <slucidi@redhat.com>