Check for deleted VMs outside of the scheduler
The check for VMs that have been deleted from the inventory
while the plan is being processed should occur outside of the
scheduler so that the scheduler implementations don't need to
be responsible for setting the `Canceled` condition on the VM.

This also fixes a bug where the `Canceled` condition was being
set on the missing VM's status, but the `NotFound` error from
the inventory was still being passed to the caller, causing
the reconcile to fail. The `NotFound` error should be set to `nil`
after the VM's status is set so that plan execution can continue.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>