MTV-2666 | VM migration with single interface from OSP fails with "more than one interface is connected to a pod network in spec.template.spec.interfaces
Issue:
When migrating an OpenStack VM that has a single interface but multiple NICs attached to it, the migration fails with an error saying there are multiple interfaces connected to the same pod.

Fix:
During destination VM creation in the mapping network phase, use only the first NIC per interface to avoid the duplication.

Ref: https://issues.redhat.com/browse/MTV-2666
Signed-off-by: Elad Hazan <ehazan@redhat.com>