Make the provider controller use new OVA setup
In order to deploy the OVA server, the provider
controller just needs to ensure that an OVAProviderServer CR
has been created in the Forklift namespace. When the
controller detects that a Provider CR has been altered
(the current generation and observed generation are different),
the OVAProviderServer CR is deleted.

When the Provider CR itself is deleted, the OVAProviderServer CR
also needs to be deleted by the controller. It can't be automatically
cleaned up because OwnerReferences cannot reach across namespaces.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>