Explicitly tear down OVA server resources
The PersistentVolume created for the OVAProviderServer
can't be automatically cleaned up by the OwnerReference
relationship since it's a cluster-scoped object. Add
a finalizer to the OVAProviderServer resource and use
that to trigger a proper cleanup process.

Also fixes the way OwnerReferences are set to make sure
they work if the TypeMeta information is missing from
objects retrieved from the client, which is apparently
a known issue with client-go for some time.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>