storage-offload: PowerMax: Use port group ID from an env var
Mapping a volume in PowerMax must contain a port-group id information.
Instead of trying to auto-detect the right port-group to use, or to
automatically create one, just specify it in the secret and that will be
consumed and used when creating the masking view.

New environment variable and secret key: POWERMAX_PORT_GROUP_ID

Signed-off-by: Roy Golan <rgolan@redhat.com>