storage-offload: PowerMax: refactor mapping handling
This change is handling 2 main areas of mapping:
1. Pass port group using env var
2. Create unique on the fly storage group, and masking view

1. Mapping a volume in PowerMax must contain a port-group id information.
Instead of trying to auto-detect the right port-group to use, or to
automatically create one, just specify it in the secret and that will be
consumed and used when creating the masking view.

New environment variable and secret key: POWERMAX_PORT_GROUP_NAME

2. Per each mapping we will create a unique storage-group and masking
   view names.
   Also, instead of using a host group we match a identity the storage
   host object by the initiators listed to it against the adapter list
   we get from vmware.

Signed-off-by: Roy Golan <rgolan@redhat.com>

storage-offload: PowerMax: unique names for storage and host groups

Signed-off-by: Roy Golan <rgolan@redhat.com>

More fixes than just a unique object name

This is the spec for the changes needed in this branch.
The change is only relevant for powermax implementation.

MaskingView changes:
create a maskingview for every population.
the maskingview name uses the xcopyInitiatorGroup variable
the masking view don't need a hostgroup, but can use the hostid directly on the masking view.
To get the host id we need to use pmax api to get the host. We know the esx hostname. hope this can help.

When unmapping the volume from that newly created masking view we need to first delete the masking view
then unmap the volume from that hostid

XcopyInitiatorGroup overriden changes:
We need to override the xcopyInitiatorGroup name, and in a generic way.
To do that change the remote_esxcli.go to check with the mappingContext if there is a variable
that is set with a name to use. That name will be set by each storage implementation in the EnsureCLonnerGroup implementation.
In powermax implementation we should set that variable to be the host.name + a short random number so that variable is unique per populatation.

Cleanup:
The cleanup in remote_esxcli.go shall have a mappingContext variable that distictivly mark it for cleanup so
in the Unmap in the storageApi we can know that this unmap is for cleanup.
Then in the Unmap implementation of powermax we need to:
- delete the masking view in Unmap impl based on mappingContext variable.
- call unmap
- delete storageGroup

- propose a clean name for the mappingContext variables
- add and/or update/delete unittest

Signed-off-by: Roy Golan <rgolan@redhat.com>

chore: storage-offload: Panic during population shall exit with an error

Panicking in the middle of the populate function must result an exit
with an error, otherwise the data is considered copied.

Signed-off-by: Roy Golan <rgolan@redhat.com>