MTV-3545 | Quick fix of cloudInitNoCloud
Issue: When migrating the VM with `cloudInitNoCloud` the migration fails
as the disk is added but the volume is missing.

Fix: This is just quick fix focused on the cloudInitNoCloud. Additional
disks like sysprep will not be migrated and for those I have created
https://github.com/kubev2v/forklift/pull/3218.
The fix adds the volume for the cloudInitNoCloud.

Signed-off-by: Martin Necas <mnecas@redhat.com>