MTV-3545 | Add missing disks in OCP to OCP migration
Issue: When migrating with cloudinit, sysprep, secrets, configmaps etc.
these need to be recreated in the target namespace.

Additional issue discovered: The destination PVCs did not have original
name but got a new plan according to the plan. This could break
customers automation.

Fix: I have added new methods to the builder interface `ConfigMaps` and
`Secrets` these will return the resources from the original VM.
Additionally I have added the Ensurer to the adapter and kubevirt which
allows us to reuse the live migratior logic for esuring the resources.
Plus this will alow us to start refactoring the original base migrator
to Ensurer.

Resolves: MTV-3545

Signed-off-by: Martin Necas <mnecas@redhat.com>