Ignore successful VMs when validating plans
VMs that have been successfully migrated and have the Succeeded status
were still validated as though they need to be migrated, which can result in conditions
on the Plan that cause it to not become Ready. This means that there are scenarios where
a Plan that partly succeeded and partly failed couldn't be re-run, because the successful
VMs are now causing spurious conditions.

We can skip validating VMs that the Plan considers to have been successfully
migrated, because they won't be remigrated by the plan and can't affect
its success. (any interference that VM would cause for another
incomplete VM on the plan would still be picked up by the other VM's
validation)

Resolves: MTV-3281 | [CCLM] "VMPowerStateUnsupported"
          and "MAC address conflicts" error after migration
          succeeded

Signed-off-by: Sam Lucidi <slucidi@redhat.com>