Handle vms that have vmexport phase of `Skipped`
When a vm does not contain any volumes that need to be exported (for
example, a vm with only ContainerDisk disks) the vmexport object will
have a phase of `Skipped`. We don't currently handle this, so MTV will
simply wait for the vmexport object to eventually reach a stage of
`Ready`, which never happens. Since we want to be able to migrate these
vms, add explicit handling for this situation.

Resolves: MTV-3516

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>