OVA: Use ResourceType to identify devices
The OVA scanner was looking at the display name of
devices in order to identify NICs and other device types.
This is incorrect and fragile, since there's no guarantee
that devices have a regular or meaningful display name.

The OVF specification provides for devices to have a
resource type which is used to identify the kind of
device and its parameters. This change set uses the
resource type to identify devices instead of display name.

Resolves: MTV-3654 | No interface related settings in VM yaml file after migrating VM from ova file to OCP cluster
Ref: https://issues.redhat.com/browse/MTV-3654
Signed-off-by: Sam Lucidi <slucidi@redhat.com>