Add retry mechanism and sleep delay to rescan function
rescan adapters comman to all devices caused an error. After the unmap a specific
storage device enters apd mode, which cause the adapter to be busy, as
it tries to reconnect to the storage device.
We'll target in the deletion the specific adapter(s) which we used, and
combine it with a retry mechanism.

Resolves: MTV-3703
Signed-off-by: Amit Weinstock <aweinsto@redhat.com>