Don't compile regex for every dynamic customization script
Currently we're looking up customization scripts and checking if they
match against a regex and returning the list of matching script names.
After we have the list of scripts, we then (for linux guests) recompile
the regex for each script and get the subgroup matches in order to
extract the action.

This patch provides a minor optimization by compiling the regex only
once and returning the match groups during the initial regex match.
We can then use these match groups later instead of recompiling and
matching in the script iteration loop.

Resolves: None

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>