Resolves: MTV-3630 | storage-offload: use leases, serialize rescans
Motivation
Storage offload and in particular adding copying on the ESXi in the
vsphere requires storage rescan, per a disk copy process. Also, VMWare
states there is a maximum cost of 2 clone operations on a single ESXi
they allow before the system will start to experience in stability.

Currently there is not mechanism to serialize or regulate the number of
concurrent requests, even with MAX_VM_INFLIGHT, because those operations
originates from the populator pod which is created by reacting to a
relevant PVC created per disk of a VM in the same stage.

Also, vib related operations like get and list, on the ESX are very
expensive and will fail quickly if called from different processes.

Modification
To orchestrate those operations and regulate them per-host we introduce
the usage of kubernetes leases form coordination.k8s.io package.
Per each host there are 2 leases named after the host and a slot. For
example for a host with managed object host-100 in VSphere, the
populator there should be up to 2 leases in the name:
esxi-lock-hostsystem-host-100-slot-0
esxi-lock-hostsystem-host-100-slot-1

Leases are exclusively under openshift-mtv namespace, so all plans use
the same leases.

When the populator code section takes a lease (again 2 are available per
host) it acquired it for 10 seconds and updates it every 3 seconds.
The lease expires after 10 seconds meaning when critical section is done
other contenders are allowed to acquire it after 10 second.
In other words the lease expires automatically.
There is no need to delete it, but other way round-the holder
needs to keep updating it.

To sync a critical section in the populator execution the remote esxcli
can now use a HostLocker inteface:

	err := hostLocker.WithLock(context.Background(), hostID,
            func() error {
                // ... critical/heavy section, e.g. rescan
            }
       )

Result
Multiple disks VMs can be served in a plan, currently tested 10 without
a problem. Looks like the heavy part of all was rescans.

Configuration:
controller_host_lease_namespace: "openshift-mtv"
controller_host_lease_duration_seconds: 10

Signed-off-by: Roy Golan <rgolan@redhat.com>