MTV-3624 | Improve authentication error logging for Ontap and PowerStore (#3671)
## Description
This PR improves the error handling logic for Ontap and PowerStore
integrations.

## Motivation
Previously, if a user provided an incorrect SVM or invalid credentials
in the backend secret, the resulting error message was generic or
misleading. This made it difficult to identify the root cause
(authentication vs. network vs. other issues).


The logs will now clearly indicate if the failure is due to
authentication/configuration issues.

Resolves: MTV-3624

---------

Signed-off-by: Amit Weinstock <aweinsto@redhat.com>