MTV-3412 | label pvc with ds volume details  (#3650)
This change adds a storage affinity hint to PVCs created during vSphere
migrations using the xcopy populator. The hint provides the NAA
identifier of the source disk to help CSI drivers make optimal placement
decisions.

## What this does

When creating PVCs for xcopy migrations, we now:
1. Query the vSphere API to retrieve the NAA ID from the VMFS datastore
2. Add a  label to the PVC with the NAA value
3. Pass this hint to the storage provider via PVC labels

## Why this matters

The NAA identifier is a vendor-neutral way to reference physical storage
devices. By providing this information, CSI drivers can:
- Create volumes on the same physical storage as the source disk
- Optimize data locality and reduce migration overhead
- Improve performance by minimizing data movement across storage arrays

This is purely advisory - CSI drivers may choose to use or ignore this
hint based on their capabilities and configuration. It's the storage
provider's responsibility to implement support for this label if
beneficial.

## Implementation details

- Added  method in vsphere/client.go to query VMFS extent info
- Created helper method in builder.go to establish vSphere connection
and retrieve NAA
- Added  constant with standardized label key
- NAA lookup integrated into  function during PVC creation
- Errors in NAA retrieval are logged but don't block PVC creation 
Resolve: MTV-3412

Signed-off-by: Amit Weinstock <aweinsto@redhat.com>