MTV-3935 | copy-offload: Refactor device detachment
Resolves: MTV-3935

MOTIVATION
When the populator is done with a LUN is unmaps it , which disconnect it from
ESX, and then rescan to remove dead devices. In some cases the rescan fails
and devices stays dead and then end in status off in the detached list.
If a restart to the pod is happening to restart the copy or the same volume is
used again then the code will show the device is recognized by the
ESX kernel (esxcli core device list -d <naa>) and we won't rescan.
If the device if status 'off' then the vmkfstools will fail,
bacause in status 'off' there is no i/o allowed. It will look in the log that
the device is found, and then the populator continues to the clone,
and then it fails.

MODIFICATION
First the cleanup should be rearranged to:
1. set the state of the device to off. then there is no more i/o; and device
goes to detached list
2. remove the device from the detached
list - esxcli storage core device detached remove -d <naa>
3. unmap - terminates the connection of the array to ESX
4. rescan to delete dead devices

Auto fixing if dead device found:
When we want to connect a device, don't continue to vmkfstools clone if the device is there.
Check if the status is 'off' . If yes then remove it from the detached list (esxcli core devices detached remove -d <naa>) and then rescan to add new devices. With that it will be in status on.

Signed-off-by: Roy Golan <rgolan@redhat.com>