MTV-3632 | Expand vsphere osmap
Resolves: MTV-3632

Issue: When migrating from VMware lot of the OS are not mapped to the
preferences.

Fix: This commit expands the osmap with all available preferences in
latest kubevirt. Additionally it adds a removal of the minor version
inside the osmap, alternative to this we would need to upadte the osmap
every time when there would be new minor release.

Signed-off-by: Martin Necas <mnecas@redhat.com>