chore(fix): make copy-offload populator settings global var
Settings are anyhow based on os.Env vars and are already being loaded.
The struct is a convenient way to get those well known values, and it
should be loaded and ready when the module loads. There is no use of
making multiple instances out of it, and/or initializing it out of the
context of the module.

Signed-off-by: Roy Golan <rgolan@redhat.com>